backend/requirements.txt
asgiref
Django==3.2.16
django-colorfield==0.8.0
django-cors-headers==3.13.0
django-filter==2.4.0
django-import-export
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields
flake8==5.0.4
gunicorn==20.0.4
Pillow==9.3.0 
psycopg2-binary==2.8.6
PyJWT==2.1.0
python-dotenv
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1